type changed

+++ 0x0E-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t int i;
 
 	if (array == NULL || size == 0 || action == NULL)
 		return;
